create `cid_hashmap` mod and modify set `insert`
