+++ src/ipld/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+mod cid_hashmap;
 mod cid_hashset;
 pub mod json;
 pub mod selector;
@@ -10,7 +11,7 @@ pub use libipld::Path;
 pub use libipld_core::ipld::Ipld;
 pub use util::*;
 
-pub use self::cid_hashset::CidHashMap;
+pub use self::cid_hashmap::CidHashMap;
 pub use self::cid_hashset::CidHashSet;
 
 pub use libipld_core::serde::{from_ipld, to_ipld};
